// enums/jobs/job-equipment-type.enum.ts
export enum JobEquipmentType {
    TRACTOR_ONLY = "TRACTOR_ONLY",
    TRACTOR_TRAILER = "TRACTOR_TRAILER",
    REEFER = "REEFER",
    FLATBED = "FLATBED",
    HEAVY_HAUL = "HEAVY_HAUL",
    HEAVY_TOWING = "HEAVY_TOWING",
    INTERMODAL = "INTERMODAL",
    AUTO_HAULING = "AUTO_HAULING",
    LOGGING = "LOGGING",
    DRY_BULK = "DRY_BULK",
    HOTSHOT = "HOTSHOT",
    DUMP_TRUCK = "DUMP_TRUCK",
    BOXTRUCK = "BOXTRUCK",
    COURIER_VAN = "COURIER_VAN",
    SERVICE_TRUCKS = "SERVICE_TRUCKS",
    DOUBLES = "DOUBLES",
    OTHER = "OTHER"
}