// enums/jobs/job-pay-method.enum.ts
export enum JobPayMethod {
    // Rate per mile
    RATE_PER_MILE = 'RATE_PER_MILE',
    // Percent per move
    PERCENT_PER_MOVE = 'PERCENT_PER_MOVE',
    // hourly
    HOURLY = 'HOURLY',
    // set weekly
    SET_WEEKLY = 'SET_WEEKLY',
    // salary
    SALARY = 'SALARY',
    // percent per weight
    PERCENT_PER_WEIGHT = 'PERCENT_PER_WEIGHT',
  }
  