// enums/jobs/job-schedule.enum.ts
export enum JobSchedule {
    MULTIPLE_WEEKS_ON_ROAD = "MULTIPLE_WEEKS_ON_ROAD",
    MOST_WEEKENDS_OFF = "MOST_WEEKENDS_OFF",
    WEEKENDS_OFF = "WEEKENDS_OFF",
    HOME_EVERY_NIGHT = "HOME_EVERY_NIGHT",
    HOME_EVERY_OTHER_NIGHT = "HOME_EVERY_OTHER_NIGHT",
    NIGHT_SHIFT = "NIGHT_SHIFT",
    OTHER = "OTHER",
}